Fix: 名前の変更、gridで並べる
